fix(native): widen the packed chain to uint64 for the eighth pedal type
Landing the wah (docs 58) and the compressor (docs 59) together makes eight
pedal types, and the APVTS board snapshot packs a 4-bit length plus kMaxChain
4-bit type slots into one atomic word. 4 x (8 + 1) = 36 bits, so the existing
static_assert fired:

    error: static assertion failed: the packed chain no longer fits a
    lock-free uint32

That assert did exactly its job - it stopped a board silently truncating
rather than letting it ship - and the comment above it had already predicted
the limit, saying there was room for a seventh slot before this needed
revisiting. The eighth is what crossed it.

Widened to uint64, still ONE lock-free word on every platform this ships to.
That property is now ASSERTED rather than assumed, via
std::atomic<juce::uint64>::is_always_lock_free, because a non-lock-free
atomic here would put a mutex on the audio thread - the one thing this
packing exists to avoid.

Native identical_core / chain_edit / cab_state all pass (3/3).

+++ native/src/PluginProcessor.h
@@ -138,7 +138,7 @@ public:
     // from 3 bits when the GOLD pedal made the board six deep: 3-bit slots would
     // still have fitted six types, but only just, and a seventh would have silently
     // aliased. Four bits keeps a whole spare type per slot and still fits a single
-    // uint32 — so the publish stays ONE relaxed atomic store, never a lock. M13.1's
+    // uint64 — so the publish stays ONE relaxed atomic store, never a lock. M13.1's
     // compressor is that seventh type, and it fits the 4-bit slots unchanged, which
     // is exactly the headroom the widening was for.
     std::vector<int> chainOrder() const;
@@ -212,7 +212,7 @@ private:
     int lastReportedLatency_ = -1;
 
     // Packed board snapshot for the audio thread (see chainOrder above).
-    std::atomic<juce::uint32> packedChain_{0};
+    std::atomic<juce::uint64> packedChain_{0};
     std::atomic<int> chainVersion_{0};
 
     // Cab-picker presentation state (message thread only).
